+++ src/main/java/f4/product/domain/product/controller/ProductController.java
@@ -5,7 +5,6 @@ import f4.product.domain.product.dto.request.ProductSaveRequestDto;
 import f4.product.domain.product.dto.request.ProductUpdateRequestDto;
 import f4.product.domain.product.dto.response.AuctionTimeStatusDto;
 import f4.product.domain.product.dto.response.FeignProductDto;
-import f4.product.domain.product.dto.response.ProductReadResponseDto;
 import f4.product.domain.product.service.AuctionStatusService;
 import f4.product.domain.product.service.ProductService;
 import f4.product.global.constant.CustomErrorCode;
@@ -23,7 +22,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
@@ -49,57 +47,6 @@ public class ProductController {
     productService.saveProduct(requestDto);
     log.info("상품 등록 완료. 상품명: {}", requestDto.getName());
     return new ResponseEntity<>("상품 등록이 완료되었습니다.", HttpStatus.OK);
-  }
-
-  @GetMapping("/findAll")
-  public ResponseEntity<List<ProductReadResponseDto>> findAll() {
-    log.info("모든 상품 조회 요청 받음.");
-    List<ProductReadResponseDto> products = productService.findAll();
-    log.info("모든 상품 조회 완료. 조회된 상품 수: {}", products.size());
-    return new ResponseEntity<>(products, HttpStatus.OK);
-  }
-
-  @GetMapping("/{productId}")
-  public ResponseEntity<ProductReadResponseDto> findById(@PathVariable Long productId) {
-    log.info("상품 id로 상품 조회 요청 받음. 상품 ID: {}", productId);
-    ProductReadResponseDto response = productService.findById(productId);
-    log.info("상품 id로 상품 조회 완료. 조회된 상품 ID: {}", response.getId());
-    return new ResponseEntity<>(response, HttpStatus.OK);
-  }
-
-  @GetMapping("/by-name")
-  public ResponseEntity<List<ProductReadResponseDto>> findByName(@RequestParam String name) {
-    log.info("상품명으로 상품 조회 요청 받음. 상품명: {}", name);
-    List<ProductReadResponseDto> products = productService.findByName(name);
-    log.info("상품명으로 상품 조회 완료. 상품명: {}, 조회된 상품 수: {}", name, products.size());
-    return new ResponseEntity<>(products, HttpStatus.OK);
-  }
-
-  @GetMapping("/by-artist")
-  public ResponseEntity<List<ProductReadResponseDto>> findByArtist(@RequestParam String artist) {
-    log.info("아티스트명으로 상품 조회 요청 받음. 아티스트명: {}", artist);
-    List<ProductReadResponseDto> products = productService.findByArtist(artist);
-    log.info("아티스트명으로 상품 조회 완료. 아티스트명: {}, 조회된 상품 수: {}", artist, products.size());
-    return new ResponseEntity<>(products, HttpStatus.OK);
-  }
-
-  @GetMapping("/search/category")
-  public ResponseEntity<List<ProductReadResponseDto>> findByMediumAndKeyword(
-      @RequestParam String category, @RequestParam String keyword) {
-    log.info("카테고리와 키워드로 상품 검색 요청 받음. 카테고리: {}, 키워드: {}", category, keyword);
-    List<ProductReadResponseDto> products =
-        productService.findByMediumAndKeyword(category, keyword);
-    log.info("카테고리와 키워드로 상품 검색 완료. 조회된 상품 수: {}", products.size());
-    return new ResponseEntity<>(products, HttpStatus.OK);
-  }
-
-  @GetMapping("/search")
-  public ResponseEntity<List<ProductReadResponseDto>> findByMedium(
-      @RequestParam String category) {
-    log.info("카테고리로 상품 검색 요청 받음. 카테고리: {}", category);
-    List<ProductReadResponseDto> products = productService.findByMedium(category);
-    log.info("카테고리로 상품 검색 완료. 조회된 상품 수: {}, 카테고리: {}", products.size(), category);
-    return new ResponseEntity<>(products, HttpStatus.OK);
   }
 
   @PutMapping("/{productId}")
